Travis is currently failing for three reasons:
1. pytest cannot automatically import pyad
2. The np power function is incorrectly specified
3. The result in the testcase base_root for derivative of y is incorrect

I changed the name of the import to pyad rather than from forward_diff import pyad. There seems to be a difference between importing within the test file (where it needs to be from forward_diff import pyad) and when running tests etc/ externally where it needs to just be import pyad.

Tests were failing with error np.pow does not exist, I changed this to np.power() which appears to have resolved the issue.

Also the test itself still failed due to the value not being correct. I checked the derivative on wolfram alpha and it looks like the value coming back from pyad is correct and the true value specified is incorrect. I updated this to match